refactor create/edit note

+++ core/exercise/src/main/kotlin/io/github/stslex/workeeper/core/exercise/data/model/ExerciseDataModel.kt
@@ -8,7 +8,7 @@ data class ExerciseDataModel(
     val name: String,
     val sets: Int,
     val reps: Int,
-    val weight: Int,
+    val weight: Double,
     val timestamp: Long,
 )
